[TSK-012] Создал интерфейс NotificationService, добавил в него методы сервиса. Изменил тип elapsed на long.

+++ src/main/java/ru/b19513/pet_schedule/repository/entity/NotificationTimeout.java
@@ -15,7 +15,7 @@ import java.util.List;
 public class NotificationTimeout extends Notification {
 
     @Column
-    private int elapsed;
+    private long elapsed;
 
     @OneToMany(fetch = FetchType.LAZY)
     private List<Period> times;
